+++ product/server/common/src/main/java/io/codekvast/common/logging/LoggingUtils.kt
@@ -59,7 +59,7 @@ object LoggingUtils {
                 .toString()
                 .substring(2)
                 .replace("(\\d[HMS])(?!$)".toRegex(), "$1 ")
-                .toLowerCase()
+                .lowercase()
     }
 
     @JvmStatic